Create CodeAlpha_Login-and-Registration-System_Syeda-Alishba-Raza
This program "Login an Registration System" is written in C++ Language. A secure and efficient user authentication system that manages user accounts through a robust login and registration process.

Key Features:

Email-based Authentication: The system checks if a user has an existing account by verifying their email address against stored files in the database.
Registration: If no account is found, users are prompted to sign up, and their information is securely stored in the database.
Password Validation: The system ensures password security by validating user input against predefined criteria.
Functionality:

Users attempt to log in with their email and password.
The system checks if an account exists by searching for a matching file in the database.
If no account is found then sends the message that email or password might be invalid try logging in again if after many tries it still fails then the user is redirected to sign up.
Upon successful registration, user information is stored in the database.
Password validity is ensured through rigorous validation checks.